+++ backend/src/main/java/peer/backend/controller/LoginController.java
@@ -5,7 +5,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import peer.backend.dto.request.UserLoginRequest;
+import peer.backend.dto.security.request.UserLoginRequest;
 import peer.backend.service.LoginService;
 
 import javax.validation.Valid;
