+++ backend/src/main/java/peer/backend/dto/security/Message.java
@@ -1,4 +1,4 @@
-package peer.backend.dto;
+package peer.backend.dto.security;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
