+++ backend/src/test/java/peer/backend/BackendApplicationTests.java
@@ -4,7 +4,7 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
-import peer.backend.dto.Message;
+import peer.backend.dto.security.Message;
 import peer.backend.service.EmailAuthService;
 import static org.assertj.core.api.Assertions.*;
